Time: N/A (0%), Space: N/A (0%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     bool isPalindrome(int x) {
         int k=x;
-        long reverse=0;
+        int reverse=0;
         while(k>0){
             int last = k%10;
             k=k/10;
